+++ neodos-kernel/src/drivers/mod.rs
@@ -1,5 +1,4 @@
 pub mod abi;
-pub mod acpi;
 pub mod ahci;
 pub mod ata;
 pub mod block;