+++ scripts/qemu-debug.sh
@@ -41,6 +41,7 @@ echo ""
 
 qemu-system-x86_64 \
   -machine q35,accel=$ACCEL \
+  -no-reboot \
   -monitor telnet:127.0.0.1:4444,server,nowait \
   -gdb tcp::1234 \
   -drive if=pflash,format=raw,readonly=on,file=$OVMF_CODE \